+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/folders/GalleryFoldersFeatureModule.kt
@@ -6,13 +6,13 @@ import org.koin.dsl.bind
 import org.koin.dsl.module
 import ua.com.radiokot.photoprism.di.APP_NO_BACKUP_PREFERENCES
 import ua.com.radiokot.photoprism.env.data.model.EnvSession
+import ua.com.radiokot.photoprism.features.albums.albumsModule
+import ua.com.radiokot.photoprism.features.albums.data.model.Album
 import ua.com.radiokot.photoprism.features.gallery.folders.data.storage.GalleryFoldersPreferences
 import ua.com.radiokot.photoprism.features.gallery.folders.data.storage.GalleryFoldersPreferencesOnPrefs
 import ua.com.radiokot.photoprism.features.gallery.folders.view.model.GalleryFolderViewModel
 import ua.com.radiokot.photoprism.features.gallery.folders.view.model.GalleryFoldersViewModel
 import ua.com.radiokot.photoprism.features.gallery.search.logic.SearchPredicates
-import ua.com.radiokot.photoprism.features.albums.data.model.Album
-import ua.com.radiokot.photoprism.features.albums.albumsModule
 
 val galleryFoldersFeatureModule = module {
     includes(albumsModule)
@@ -31,10 +31,15 @@ val galleryFoldersFeatureModule = module {
                 albumsRepository = get(),
                 preferences = get(),
                 searchPredicate = { album: Album, query: String ->
-                    val pathFragments = album.path
-                        .split('/')
-                        .filterNot(String::isEmpty)
-                    SearchPredicates.generalCondition(query, pathFragments + album.title)
+                    val fields = buildList {
+                        add(album.title)
+                        album.path
+                            ?.split('/')
+                            ?.filterNot(String::isEmpty)
+                            ?.also(::addAll)
+                    }
+
+                    SearchPredicates.generalCondition(query, fields)
                 },
             )
         }